+++ datarec/io/writers/sequences/__init__.py
@@ -1,2 +1,2 @@
-from .json import write_sequences_json, write_sequences_json_array
+from .json import write_sequences_json, write_sequences_json_array, write_sequences_json_items
 from .tabular import write_sequence_tabular_inline, write_sequence_tabular_wide, write_sequence_tabular_implicit