+++ examples/in_out/writers/sequences.py
@@ -2,7 +2,7 @@ import os
 from datarec import DataRec
 from datarec.io.readers.transactions.tabular import read_transactions_tabular
 from datarec.io.writers.sequences.tabular import write_sequence_tabular_inline, write_sequence_tabular_wide, write_sequence_tabular_implicit
-from datarec.io.writers.sequences.json import write_sequences_json, write_sequences_json_array
+from datarec.io.writers.sequences.json import write_sequences_json, write_sequences_json_array, write_sequences_json_items
 from datarec.datasets.examples import download_example_by_url
 
 output_folder = 'data'
@@ -22,6 +22,7 @@ write_sequence_tabular_inline(rawdata, os.path.join(output_folder, 'sequence_inl
 write_sequence_tabular_wide(rawdata, os.path.join(output_folder, 'sequence_wide.tsv'), user_col='user', col_sep=',', header=True, verbose=True)
 write_sequence_tabular_implicit(rawdata, os.path.join(output_folder, 'sequence_implicit.tsv'), col_sep=',', verbose=True)
 write_sequences_json(rawdata, os.path.join(output_folder, 'sequence.json'), include_rating=True, include_timestamp=True, item_col='item', rating_col='rating', timestamp_col='timestamp', verbose=True)
+write_sequences_json_items(rawdata, os.path.join(output_folder, 'sequence_items.json'), item_col='item', verbose=True)
 write_sequences_json_array(rawdata, os.path.join(output_folder, 'sequence_array.json'), include_rating=True, include_timestamp=True, user_col='user', item_col='item', rating_col='rating', timestamp_col='timestamp', verbose=True)
 
 datarec = DataRec(rawdata=rawdata)
@@ -29,4 +30,5 @@ write_sequence_tabular_inline(datarec, os.path.join(output_folder, 'sequence_inl
 write_sequence_tabular_wide(datarec, os.path.join(output_folder, 'sequence_wide.tsv'), user_col='user', col_sep=',', header=True, verbose=True)
 write_sequence_tabular_implicit(datarec, os.path.join(output_folder, 'sequence_implicit.tsv'), col_sep=',', verbose=True)
 write_sequences_json(datarec, os.path.join(output_folder, 'sequence.json'), include_rating=True, include_timestamp=True, item_col='item', rating_col='rating', timestamp_col='timestamp', verbose=True)
+write_sequences_json_items(datarec, os.path.join(output_folder, 'sequence_items.json'), item_col='item', verbose=True)
 write_sequences_json_array(datarec, os.path.join(output_folder, 'sequence_array.json'), include_rating=True, include_timestamp=True, user_col='user', item_col='item', rating_col='rating', timestamp_col='timestamp', verbose=True)